+++ src/javascript/top_bar.js
@@ -279,4 +279,3 @@ graviton.loadControlButtons = ()=>{
     document.getElementById("controls").setAttribute("os", "not_windows");
   }
 }
-